Factor data model concerns from Location module -> Data module.
