+++ Sources/Data/Model/Beacon.swift
@@ -7,7 +7,6 @@
 //
 
 import CoreData
-import CoreLocation
 import os.log
 
 public final class Beacon: NSManagedObject {
@@ -24,15 +23,21 @@ public final class Beacon: NSManagedObject {
 
     @NSManaged public private(set) var regionIdentifier: String
     
+
+    
     public override func awakeFromInsert() {
         super.awakeFromInsert()
         self.tags = []
     }
     
     public override func willSave() {
-        if self.regionIdentifier != self.region.identifier {
-            self.regionIdentifier = self.region.identifier
+        if self.regionIdentifier != generateIdentifier() {
+            self.regionIdentifier = generateIdentifier()
         }
+    }
+    
+    func generateIdentifier() -> String {
+        return "\(self.uuid.uuidString):\(self.major):\(self.minor)"
     }
 }
 
@@ -50,19 +55,6 @@ extension Beacon: AttributeRepresentable {
         ]
         
         return .object(attributes)
-    }
-}
-
-// MARK: Core Location
-
-extension Beacon {
-    public var region: CLBeaconRegion {
-        return CLBeaconRegion(
-            proximityUUID: self.uuid,
-            major: UInt16(self.major),
-            minor: UInt16(self.minor),
-            identifier: "\(self.uuid.uuidString):\(self.major):\(self.minor)"
-        )
     }
 }
 
@@ -124,22 +116,5 @@ extension Beacon {
         } catch {
             os_log("Failed to delete beacons: %@", log: .persistence, type: .error, error.localizedDescription)
         }
-    }
-}
-
-// MARK: Collection
-
-extension Collection where Element == Beacon {
-    public func wildCardRegions(maxLength: Int) -> Set<CLBeaconRegion> {
-        let uuids = self.map({ $0.uuid })
-        let unique = Set(uuids)
-        
-        #if swift(>=4.2)
-        let regions = unique.shuffled().prefix(maxLength).map { $0.region }
-        #else
-        let regions = unique.prefix(maxLength).map { $0.region }
-        #endif
-        
-        return Set<CLBeaconRegion>(regions)
     }
 }
