+++ Sources/Data/Model/Geofence.swift
@@ -7,7 +7,6 @@
 //
 
 import CoreData
-import CoreLocation
 import os
 
 public final class Geofence: NSManagedObject {
@@ -30,9 +29,13 @@ public final class Geofence: NSManagedObject {
     }
     
     public override func willSave() {
-        if self.regionIdentifier != self.region.identifier {
-            self.regionIdentifier = self.region.identifier
+        if self.regionIdentifier != self.generateIdentifier() {
+            self.regionIdentifier = self.generateIdentifier()
         }
+    }
+    
+    public func generateIdentifier() -> String {
+        return "\(latitude):\(longitude):\(radius)"
     }
 }
 
@@ -47,26 +50,6 @@ extension Geofence: AttributeRepresentable {
             "radius": self.radius,
             "tags": self.tags
         ]
-    }
-}
-
-// MARK: Core Location
-
-extension Geofence {
-    public var coordinate: CLLocationCoordinate2D {
-        return CLLocationCoordinate2D(latitude: latitude, longitude: longitude)
-    }
-    
-    public var location: CLLocation {
-        return CLLocation(latitude: latitude, longitude: longitude)
-    }
-    
-    public var region: CLCircularRegion {
-        return CLCircularRegion(
-            center: location.coordinate,
-            radius: radius,
-            identifier: "\(latitude):\(longitude):\(radius)"
-        )
     }
 }
 
@@ -167,47 +150,5 @@ extension Geofence {
         } catch {
             os_log("Failed to delete geofences: %@", log: .persistence, type: .error, error.localizedDescription)
         }
-    }
-}
-
-// MARK: Collection
-
-extension Collection where Element == Geofence {
-    public func sortedByDistance(from coordinate: CLLocationCoordinate2D) -> [Geofence] {
-        os_log("Sorting geofences...", log: .general, type: .debug)
-        
-        #if swift(>=4.2)
-        if #available(iOS 12.0, *) {
-            os_signpost(.begin, log: .general, name: "sortGeofences")
-        }
-        #endif
-        
-        let sorted = self.sorted(by: {
-            return coordinate.distanceTo($0.coordinate) < coordinate.distanceTo($1.coordinate)
-        })
-        
-        #if swift(>=4.2)
-        if #available(iOS 12.0, *) {
-            os_signpost(.end, log: .general, name: "sortGeofences")
-        }
-        #endif
-        
-        os_log("Sorted %d geofences", log: .general, type: .debug, self.count)
-        return sorted
-    }
-    
-    public func regions(closestTo coordinate: CLLocationCoordinate2D?, maxLength: Int) -> Set<CLCircularRegion> {
-        let regions: [CLCircularRegion]
-        if let coordinate = coordinate {
-            regions = self.sortedByDistance(from: coordinate).prefix(maxLength).map { $0.region }
-        } else {
-            #if swift(>=4.2)
-            regions = self.shuffled().prefix(maxLength).compactMap { $0.region }
-            #else
-            regions = self.prefix(maxLength).compactMap { $0.region }
-            #endif
-        }
-        
-        return Set<CLCircularRegion>(regions)
     }
 }
