+++ Sources/Data/Model/Rover.xcdatamodeld/RoverLocation.xcdatamodel/contents
